Moved transformations.js to helpers folder
